Renamed LimitOrderLike and MarketOrderLike traits.
