+++ src/main/scala-2.11/markets/orders/AskOrder.scala
@@ -32,8 +32,10 @@ trait AskOrder extends Order {
 
   /** Splits an existing `AskOrder` into two separate orders.
     *
-    * @param residualQuantity
+    * @param residualQuantity the quantity of the residual, unfilled portion of the `AskOrder`.
     * @return a tuple of ask orders.
+    * @note The first order in the tuple represents the filled portion of the `AskOrder`; the
+    *       second order in the tuple represents the residual, unfilled portion of the `AskOrder`.
     */
   def split(residualQuantity: Long): (AskOrder, AskOrder)
